Move failedIngestJob to IngestJobStatusFromJobTestData
