+++ java/ingest/ingest-core/src/test/java/sleeper/ingest/core/task/IngestTaskTest.java
@@ -49,8 +49,8 @@ import static sleeper.core.tracker.ingest.job.IngestJobUpdateType.STARTED;
 import static sleeper.ingest.core.IngestResultTestData.defaultFileIngestResult;
 import static sleeper.ingest.core.IngestResultTestData.defaultFileIngestResultReadAndWritten;
 import static sleeper.ingest.core.job.IngestJobTestData.DEFAULT_TABLE_ID;
+import static sleeper.ingest.core.job.status.IngestJobStatusFromJobTestData.failedIngestJob;
 import static sleeper.ingest.core.job.status.IngestJobStatusFromJobTestData.finishedIngestJobUncommitted;
-import static sleeper.ingest.core.job.status.IngestJobStatusTestHelper.failedIngestJob;
 import static sleeper.ingest.core.task.IngestTaskStatusTestData.finishedMultipleJobs;
 import static sleeper.ingest.core.task.IngestTaskStatusTestData.finishedNoJobs;
 import static sleeper.ingest.core.task.IngestTaskStatusTestData.finishedOneJob;
